python_server/requirements.txt
flask
openai
flask-sqlalchemy
nltk
scikit-learn
newsapi-python
flask-apscheduler
flask-cors
python-dotenv
gunicorn